feat(index.html): enhance SEO and social media sharing metadata feat(public): add new og-image.png for social media sharing
The changes in the index.html file are aimed at improving the SEO of the website. The title and description have been updated to be more descriptive and keyword-rich. Additional meta tags have been added for social media sharing on platforms like Facebook (og tags) and Twitter. These changes will help improve the visibility of the website on search engines and make the shared links on social media more attractive.

A new image (og-image.png) has been added to the public directory. This image is referenced in the meta tags and will be used as the preview image when the website link is shared on social media platforms. This will make the shared links more visually appealing and potentially increase click-through rates.