feat(EmbedModal.tsx): remove Twitter embed provider feat(Cursors.tsx): add label to MouseState and display Moodie component inside Cursor refactor(Cursor.tsx): allow children props for custom content inside Cursor
The Twitter embed provider was removed from the EmbedModal.tsx file due to the lack of support for Twitter embeds in the current version of the application.

In Cursors.tsx, a label property was added to the MouseState type to store the base58 encoded public key of the client. This label is then used as the name for the Moodie component, which is now displayed inside the Cursor component. This change enhances the user experience by providing a visual representation of the client's public key.

The Cursor component was refactored to accept children props. This allows for custom content to be displayed inside the Cursor, providing more flexibility for future enhancements or customizations.